+++ src/utils.rs
@@ -25,10 +25,6 @@ pub fn escape_html(input: &str) -> String {
             '"' => output.push_str("&quot;"),
             '\'' => output.push_str("&#x27;"),
             '/' => output.push_str("&#x2F;"),
-            // Additional one for old IE (unpatched IE8 and below)
-            // TODO: do we still care about that? unpatched <=IE8, c'mon
-            // See https://github.com/OWASP/owasp-java-encoder/wiki/Grave-Accent-Issue
-            '`' => output.push_str("&#96;"),
             _ => output.push(c),
         }
     }