+++ CHANGELOG.md
@@ -13,7 +13,7 @@ a `Context` by value to avoid cloning and `render_value/one_off_value` takes som
 `Serialize`
 - Require macros import to be at the top of the files along with `extends` as it is fairly cheap and
 the code already only really look there.
-
+- Enforce spacing in tags at the parser, before `ifsomething` was considered ok
 
 ### Others