+++ CHANGELOG.md
@@ -29,7 +29,7 @@ only at the last position
 - Add `Context::from_serialize` to instantiate a `Context` from something that impl `Serialize`
 - Make tests helper fns `number_args_allowed`, `value_defined` and `extract_string` public
 - Add `else` clause to for loops
-- Filters are now evaluated when checking if/elif conditins
+- Filters are now evaluated when checking if/elif conditions
 
 ## 0.11.20 (2018-11-14)