+++ CHANGELOG.md
@@ -31,6 +31,8 @@ only at the last position
 - Add `else` clause to for loops
 - Filters are now evaluated when checking if/elif conditions
 - Allow `{{-` and `-}}` for whitespace management
+- Adds a `xml_escape` filter
+- Grave accent is no longer escaped in HTML, not really needed anymore
 
 ## 0.11.20 (2018-11-14)